test(integration): add test for anonymouse session flow
